+++ libs/core/src/state/settings/by_widget.rs
@@ -35,7 +35,11 @@ impl SettingsByWidget {
             "@seelen/window-manager" => self.wm.enabled,
             "@seelen/wallpaper-manager" => self.wall.enabled,
             "@seelen/launcher" => self.launcher.enabled,
-            _ => self.others.get(widget_id).is_some_and(|s| s.enabled),
+            _ => match self.others.get(widget_id) {
+                Some(settings) => settings.enabled,
+                // only official widgets are enabled by default
+                None => widget_id.starts_with("@seelen"),
+            },
         }
     }